+++ test/core/string_type.sv
@@ -1,0 +1,2 @@
+`define TYPE string
+`include "string_type.vh"