+++ test/core/string_type.v
@@ -1,0 +1,2 @@
+`define TYPE
+`include "string_type.vh"